+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+
+## [2.25.0] - 2026-03-25
 ### Added
 - holderId and softDescriptor attributes to MerchantSession and MerchantPurchase